Split web launcher provider contract modules
