+++ .pm/tasks/task_169255fb26a2410a9c9edfaa839fc466.execution.md
@@ -428,3 +428,34 @@ Example:
 - Expected Result: Branch is pushed and a GitHub PR is created with task evidence in the body.
 - Actual Result: prepare-task-pr.sh passed preflight and pushed branch, then gh pr create failed because body/fill was missing in non-interactive mode. Direct gh pr create fallback succeeded: PR URL https://github.com/eng-cc/oasis7/pull/482.
 - Blocker / Next Action: No PR creation blocker remains; continue post-PR watch/fix/merge flow.
+
+## 2026-06-15 23:34:00 CST / viewer_engineer
+- 完成内容: Implemented narrow PR #482 required-gate file-size fix for Launcher/Web Rust files without behavior changes.
+- Attribution Boundary: This is a `viewer_engineer` implementation slice for the Rust file-size gate only. It preserves launcher behavior by moving existing provider config/arg helper code and provider contract tests into sibling module files; no unrelated main merge content was intentionally changed.
+- Changed Files:
+  - `crates/oasis7/src/bin/oasis7_web_launcher/control_plane.rs`
+  - `crates/oasis7/src/bin/oasis7_web_launcher/control_plane/provider_config.rs`
+  - `crates/oasis7/src/bin/oasis7_web_launcher/oasis7_web_launcher_tests.rs`
+  - `crates/oasis7/src/bin/oasis7_web_launcher/provider_contract_tests.rs`
+- Design Choice:
+  - Moved web launcher provider arg-building/canonicalization/validation helpers behind `control_plane/provider_config.rs`, leaving `control_plane.rs` with a small `collect_agent_provider_config_issues` call and existing public behavior.
+  - Moved agent-provider contract tests into `provider_contract_tests.rs` and kept the parent test module declaration so Rust test discovery still finds the same test functions under a submodule.
+- Validation Command: `limit=1200; for f in crates/oasis7/src/bin/oasis7_web_launcher/control_plane.rs crates/oasis7/src/bin/oasis7_web_launcher/oasis7_web_launcher_tests.rs crates/oasis7/src/bin/oasis7_web_launcher/control_plane/provider_config.rs crates/oasis7/src/bin/oasis7_web_launcher/provider_contract_tests.rs; do n=$(wc -l < "$f" | tr -d ' '); echo "$n $f"; test "$n" -lt "$limit" || exit 1; done`
+- Expected Result: All changed/split Rust files are below the 1200-line gate.
+- Actual Result: passed; line counts were 1087, 1164, 336, and 234 respectively.
+- Validation Command: `cargo fmt --check && git diff --check`
+- Expected Result: Formatting and whitespace checks pass.
+- Actual Result: passed, exit code 0.
+- Validation Command: `./scripts/cargo-dev.sh test -p oasis7 --bin oasis7_web_launcher web_launcher_rejects -- --nocapture`
+- Expected Result: Focused web launcher rejection tests compile and pass.
+- Actual Result: not completed; command waited on the shared Cargo build directory lock for about 90 seconds and was interrupted with Ctrl-C, exit code 130. No Rust test pass/fail result claimed.
+- Blocker / Next Action: Ensure the two new untracked split files are staged with the PR fix, then rerun `./scripts/check-rust-file-size.sh` after staging/commit so git-ls-files includes them; rerun the focused Rust test when the shared Cargo target lock clears or rely on GitHub CI for the compile gate.
+
+## 2026-06-15 23:48:41 CST / tpm
+- 完成内容: Addressed PR #482 required-gate failure from check-rust-file-size by integrating viewer_engineer slice: split web launcher provider config helpers into control_plane/provider_config.rs and provider contract tests into provider_contract_tests.rs; adjusted test module wiring at crate root.
+- 遗留事项: Push CI fix and watch required-gate rerun; merge only after required checks pass and comments/reviews remain non-blocking.
+- Action: Commit and push file-size split fix; rerun GitHub PR checks.
+- Validation Command: cargo fmt --check; ./scripts/check-rust-file-size.sh; git diff --check; env -u RUSTC_WRAPPER cargo test -p oasis7 --bin oasis7_web_launcher web_launcher_config -- --list
+- Expected Result: Formatting, file-size gate, and whitespace pass; test discovery confirms split provider contract tests are visible.
+- Actual Result: cargo fmt --check passed; ./scripts/check-rust-file-size.sh passed with oversized code files=0, test files=0, limit=1200; git diff --check passed. Local cargo test/list attempts were repeatedly blocked by shared Cargo build/package locks and were interrupted without pass/fail after the module wiring change; GitHub required-gate must cover compile/test discovery.
+- Blocker / Next Action: No local file-size blocker remains; push the fix and continue PR CI watch.
